refactor: 질답페이지 커스텀훅으로 분리 #33
